AromaLux/requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.0.1
MySQLdb==2.2.0
Werkzeug==2.3.7
Faker==19.6.2
python-dotenv==1.0.0
Flask-CORS==4.0.0
WTForms==3.0.1
email-validator==2.0.0
gunicorn==21.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
